django_react_project/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
black==19.10b0
Click==7.0
Django==3.0
django-cors-headers==3.2.0
django-webpack-loader==0.6.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.6.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
regex==2019.12.9
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
wrapt==1.11.2
